+++ src/raycaster/math_utils.c
@@ -6,13 +6,13 @@
 /*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/21 10:29:33 by abied-ch          #+#    #+#             */
-/*   Updated: 2024/01/23 23:46:52 by abied-ch         ###   ########.fr       */
+/*   Updated: 2024/01/25 12:45:05 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/raycast.h"
 
-bool	collision(t_data *data, float new_x, float new_y)
+bool	collision(t_data *data, double new_x, double new_y)
 {
 	int	map_pos;
 
@@ -28,7 +28,7 @@ bool	collision(t_data *data, float new_x, float new_y)
 	return (true);
 }
 
-void	normalize_angle(float *angle, float increment)
+void	normalize_angle(double *angle, double increment)
 {
 	*angle += increment;
 	if (*angle < 0)
@@ -37,7 +37,7 @@ void	normalize_angle(float *angle, float increment)
 		*angle -= 2 * PI;
 }
 
-float	dist(t_line line)
+double	dist(t_line line)
 {
 	return (sqrt((line.x1 - line.x0) * (line.x1 - line.x0)
 			+ (line.y1 - line.y0) * (line.y1 - line.y0)));
